requirement.txt
surya-ocr==0.14.0
pdf2image==1.17.0
PyPDF2==3.0.1
python-doctr==1.0.0